add ESLint clean up a few things archive old init file add SHPS bootstrap
