+++ .gitattributes
@@ -1,7 +1,7 @@
 ###############################################################################
 # Set default behavior to automatically normalize line endings.
 ###############################################################################
-* text=auto
+* text eol=lf
 
 ###############################################################################
 # Set default behavior for command prompt diff.
