+++ .gitignore
@@ -1,1 +1,5 @@
-config/config.json.bak
+# NodeJS
+node_modules
+
+# JetBrains
+.idea/
