+++ package.json
@@ -1,51 +1,55 @@
 {
-    "name": "SHPS",
-    "version": "4.4.2",
-    "description": "SHPS",
-    "main": "SHPS.js",
-    "homepage": "https://shps.io",
-    "author": {
-        "name": "Marco Alka",
-        "email": "me@marco-alka.de",
-        "url": "https://marco-alka.de"
-    },
-	"repository": {
-		"type" : "git",
-		"url" : "https://github.com/Skellods-Network/SHPS4Node.git"
-	},
-    "dependencies": {
-        "vasync": "^1.6.0",
-        "colors": ">=0.6.0 < 2",
-        "http2": ">=2.7.0 < 4",
-        "lodash": ">=3.0.0 < 5",
-        "q": "^1.2.0",
-        "object-assign": ">=2.0.0 < 5",
-        "request-ip": "^1.1.4",
-        "node-mb-string-size": "^1.0.0",
-        "node-mod-load": "^2.1.0",
-        "promise-defer": "^1.0.0",
-        "busboy": "^0.2.0",
-        "qs": "^6.0.0",
-        "SHPS4Node-auth": "1.0.x",
-        "SHPS4Node-cache": "1.0.x",
-        "SHPS4Node-commandline": "1.1.x",
-        "SHPS4Node-config": "1.1.x",
-        "SHPS4Node-cookie": "1.0.x",
-        "SHPS4Node-CSS": "1.1.x",
-        "SHPS4Node-dependency": "1.0.x",
-        "SHPS4Node-error": "1.0.x",
-        "SHPS4Node-file": "1.1.x",
-        "SHPS4Node-frontend": "1.0.x",
-        "SHPS4Node-init": "1.0.x",
-        "SHPS4Node-language": "1.0.x",
-        "SHPS4Node-log": "1.0.x",
-        "SHPS4Node-make": "1.0.x",
-        "SHPS4Node-optimize": "1.0.x",
-        "SHPS4Node-parallel": "1.0.x",
-        "SHPS4Node-plugin": "1.0.x",
-        "SHPS4Node-sandbox": "1.1.x",
-        "SHPS4Node-schedule": "1.0.x",
-        "SHPS4Node-session": "1.0.x",
-        "SHPS4Node-SQL": "1.1.x"
-    }
+  "name": "SHPS",
+  "version": "5.0.0",
+  "cycle": "X",
+  "description": "SHPS",
+  "license": "MPL-2.0",
+  "main": "SHPS.js",
+  "homepage": "https://shps.io",
+  "author": {
+    "name": "Marco Alka",
+    "email": "me@marco-alka.de",
+    "url": "https://marco-alka.de"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Skellods-Network/SHPS4Node.git"
+  },
+  "dependencies": {
+    "vasync": "^1.6.0",
+    "colors": ">=0.6.0 < 2",
+    "http2": ">=2.7.0 < 4",
+    "lodash": ">=3.0.0 < 5",
+    "request-ip": "^1.1.4",
+    "node-mb-string-size": "^1.0.0",
+    "node-mod-load": "^3.1.0",
+    "promise-defer": "^1.0.0",
+    "busboy": "^0.2.0",
+    "qs": "^6.0.0",
+    "SHPS4Node-auth": "^1.0.0",
+    "SHPS4Node-cache": "^1.0.0",
+    "SHPS4Node-commandline": "^1.1.0",
+    "SHPS4Node-config": "^1.1.0",
+    "SHPS4Node-cookie": "^1.0.0",
+    "SHPS4Node-CSS": "^1.1.0",
+    "SHPS4Node-dependency": "^1.0.0",
+    "SHPS4Node-error": "^1.0.0",
+    "SHPS4Node-file": "^1.1.0",
+    "SHPS4Node-frontend": "^1.0.0",
+    "SHPS4Node-init": "^1.0.0",
+    "SHPS4Node-language": "^1.0.0",
+    "SHPS4Node-log": "^1.0.0",
+    "SHPS4Node-make": "^1.0.0",
+    "SHPS4Node-optimize": "^1.0.0",
+    "SHPS4Node-parallel": "^1.0.0",
+    "SHPS4Node-plugin": "^1.0.0",
+    "SHPS4Node-sandbox": "^1.1.0",
+    "SHPS4Node-schedule": "^1.0.0",
+    "SHPS4Node-session": "^1.0.0",
+    "SHPS4Node-SQL": "^1.1.0"
+  },
+  "devDependencies": {
+    "eslint": "^3.13.0",
+    "eslint-config-google": "^0.7.0"
+  }
 }
